# Eliminate `EGraph.getClasses` method