This PR eliminates the expensive `EGraph.getClasses`, in favor of a more efficient and more convenient `EGraph.getAllApplications` method.

This gives a noticeable performance improvement:

* [current main](https://github.com/CurtisFenner/shiru-ts/actions/runs/4707233861/jobs/8348985363)
    * ```
      Slowest: verify_tests.duplicated-assert-is-fast took 3468 ms
      Done in 6.03s.
        ```
* [this commit](https://github.com/CurtisFenner/shiru-ts/actions/runs/4821745082/jobs/8588037108)
    * ```
      Slowest: verify_tests.duplicated-assert-is-fast took 2672 ms
      Done in 4.94s.
      ```